feat: implementar módulo AlimentoComida completo con CRUD y validaciones
Módulo Rutina completo:
- 2 DTOs
- Corrección de tipos: cantidadGramos de Integer a BigDecimal
- AlimentoComidaMapper con MapStruct
- IAlimentoComidaRepository con 8 métodos de consulta
- IAlimentoRepository e IComidRepository inyectados para validar relaciones
- IAlimentoComidaService con 13 métodos
- AlimentoComidaService con validaciones, transacciones y logs
- AlimentoComidaController con 15 endpoints REST documentados

Correcciones een DTOs:
- AlimentoComidaDTO: cantidadGramos BigDecimal (era Integer)
- AlimentoCommidaCreateDTO: cantidadGramos BigDecimal (era Integer)

Todo comprobado y funcionando correctamente.

+++ api/gymprofit-api/src/main/java/com/gymprofit/api/dto/entity/alimentocomida/AlimentoComidaDTO.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -13,6 +14,6 @@ public class AlimentoComidaDTO implements Serializable {
     private Integer id;
     private Integer comidaId;
     private Integer alimentoId;
-    private Integer cantidadGramos;
+    private BigDecimal cantidadGramos;
     private Integer caloriasTotales;
 }
